add restriction

+++ app/controllers/LateDeductController.php
@@ -19,6 +19,10 @@ class LateDeductController
 
     public function store()
     {
+        if (Auth::user('role_id') != 1) {
+            redirect('home');
+        }
+
         $request = Request::validate('late/deduction', [
             'deduct-amount' => 'required'
         ]);
@@ -42,6 +46,10 @@ class LateDeductController
 
     public function entry()
     {
+        if (Auth::user('role_id') != 1) {
+            redirect('home');
+        }
+
         $pageTitle = "Late Entry";
 
         $users_data = App::get('database')->selectLoop('*', 'users');
@@ -51,6 +59,10 @@ class LateDeductController
 
     public function newLate()
     {
+        if (Auth::user('role_id') != 1) {
+            redirect('home');
+        }
+
         $deduction_amount = App::get('database')->select('amount', 'late_deduction', 'id > 0 ORDER BY id DESC LIMIT 1');
 
         $request = Request::validate('late/entry', [
@@ -70,12 +82,20 @@ class LateDeductController
 
     public function delete($id)
     {
+        if (Auth::user('role_id') != 1) {
+            redirect('home');
+        }
+
         App::get('database')->delete('late_history', "id = '$id'");
         redirect('entry', "success delete.");
     }
 
     public function payment()
     {
+        if (Auth::user('role_id') != 1) {
+            redirect('home');
+        }
+
         $pageTitle = "Late Payment";
         $users_data = App::get('database')->selectLoop('*', 'users');
         $payment_datas = App::get('database')->selectLoop('*', 'payment', 'id > 0 ORDER BY id DESC');
@@ -84,6 +104,10 @@ class LateDeductController
 
     public function storePayment()
     {
+        if (Auth::user('role_id') != 1) {
+            redirect('home');
+        }
+
         $request = Request::validate('late/payment', [
             'pay_user' => 'required',
             'pay_amount' => 'required'
@@ -101,6 +125,10 @@ class LateDeductController
 
     public function deletePayment($id)
     {
+        if (Auth::user('role_id') != 1) {
+            redirect('home');
+        }
+
         App::get('database')->delete('payment', "id = '$id'");
         redirect('payment', "success delete.");
     }
